Refactor compaction and harden planning writes
